+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Verison 0.0.2
 
 - Introduce s2.RegionCoverer.getRadiusCovering methods that take a lat/lng and radius and generates a covering
+- Expose the Point & Cell class and Cell#getVertex method
 
 ## Version 0.0.1